Add scripts to backup and restore timestamps
Using the timestamps of the files on disk was in hindsight not the best
design decision for qb. First, in order to have a creation timestamp,
the ugly `.cre` file hack had to be introduced. Second, version control
systems like Git mess the filestamps up as well (with good intentions,
though).

Since I'm moving away from using qb in my personal blog, I want to
create a long-term archive of my posts, one which does not depend on me
not losing the mtimes during a copy operation. Therefore, these two
little scripts help me capture the timestamps as they are on the host my
blog ran on since 2013. I'm gonna commit the resulting file to Git to
keep me from losing the timestamps ever again.

This would be tiresome if I was still writing new blog posts, but I don't.